[Improvement-6889] split the routing module to increase subsequent maintainability.
