+++ dolphinscheduler-ui/src/js/conf/home/router/module/user.js
@@ -1,0 +1,60 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import i18n from '@/module/i18n/index.js'
+
+const user = [
+  {
+    path: '/user',
+    name: 'user',
+    component: resolve => require(['../../pages/user'], resolve),
+    meta: {
+      title: `${i18n.$t('User Center')}`
+    },
+    redirect: {
+      name: 'account'
+    },
+    children: [
+      {
+        path: '/user/account',
+        name: 'account',
+        component: resolve => require(['../../pages/user/pages/account'], resolve),
+        meta: {
+          title: `${i18n.$t('User Information')}`
+        }
+      },
+      {
+        path: '/user/password',
+        name: 'password',
+        component: resolve => require(['../../pages/user/pages/password'], resolve),
+        meta: {
+          title: `${i18n.$t('Edit password')}`
+        }
+      },
+      {
+        path: '/user/token',
+        name: 'token',
+        component: resolve => require(['../../pages/user/pages/token'], resolve),
+        meta: {
+          title: `${i18n.$t('Token manage')}`
+        }
+      }
+    ]
+  }
+]
+
+export default user
